Set timezone data in a platform independent way (#72)
Mounting /etc/timezone is not correct across platforms and can lead
to corruption of the timezone database.  This is why Docker does not
permit the timezone file to be mounted into the container.  Instead,
pass an Environment variable containing `TZ`. Most modern apps
understand how to process `TZ`.  For those that don't, I plan to
introduce an entrypoint to the docker command to overwrite /etc/localtime
with a new file that contains the contents of this environment variable.