Address reviewer comments.
Use IN_CONTAINER_BUILD to indicate to Makefile.core.mk that we
are within a container build.  This is necessary for some use cases.
If make is recursively called without using `-f Makefile.core.mk` this
environment value may be lost.  I am not certain as I didn't test that
recursion scenario.