Pass BUILD_WITH_CONTAINER to the child containers (#91)
* WIP Pass BUILD_WITH_CONTAINER to the child containers

The per-repo makefiles environment in some cases needs to know if they
are being built in the build-container.  To provide that information, pass
it in via the environment.

Depends-On: https://github.com/istio/tools/pull/402

* Address reviewer comments.

Use IN_CONTAINER_BUILD to indicate to Makefile.core.mk that we
are within a container build.  This is necessary for some use cases.
If make is recursively called without using `-f Makefile.core.mk` this
environment value may be lost.  I am not certain as I didn't test that
recursion scenario.

* Use build tools: 2019-10-02T14-57-08