Relax consistency in the development environment
For dockerd to maintain consistency in teh bind mounts on Linux machines,
its relatively easy to achieve this result just by using native Linux
interfaces.

On MacOS, Docker has to do **ALOT** more work to maintain filesystem
consistency. I use the following definition: A write to a shared mount
from either the development environment or within the container will be
 read exactly as it was written, instead of potentially returning old data
on a read.

If the consistency is relaxed (I am talking a few hundred milliseconds at
most), Docker for Mac performs much better with minimal to no noticable
impact. I would like to trial this PR in 1.7. If it becomes a problem,
it is easy enough to revert.

Some (unofficial) benchmarks:

 Freshly installed MacOS on 2020 Mac 13" with 32gb 3733 LPDDR4, 1TB hd, 2.3 GHZ Quad Core Intel core I7 lightly loaded running on wall power.

Docker is configured as:
6 CPUs, 8GB ram, 1gig swap setup for Docker for  Mac, running K8s

| with PR | cached | benchmark command | wall-clock (s) |
| --- | --- | --- | --- |
| YES | NO | gtime -f %e make clean | 1.53
| YES | NO | gtime -f %e make build | 48.54
| YES | NO | gtime -f %e make docker | 68.49
| YES | YES | gtime -f %e make build | 9.45
| YES | YES | gtime -f %e make docker | 40.04
| NO | NO | gtime -f %e make clean | 2.23
| NO | NO | gtime -f %e make build | 53.20
| NO | NO | gtime -f %e make docker | 91.86
| NO | YES | gtime -f %e make build | 23.07
| NO | YES | gtime -f %e make docker | 59.15