Use /home/.cache as a volume mount (#284)
* Use /home/.cache as a volume mount

The proxy build uses /home/.cache to store envoy proxy output. Use that
directory as a volume mount to store the output of the build results.
This takes a 20 minute build to 12 seconds on my gear. This does not
solve the problem that the proxy build does not output binaries,
however, it gets Istio closer to that goal.

* Use same consistency mode as other volume mounts