KinD: Fix access to loadbalancerIPs from the host network
When deploying a kind cluster using integ-suite-kind.sh along
with metallb, it was seen that LoadBalancer IPs were inaccessible
from the host network. This is because kubeadm clusters automatically
configure the node.kubernetes.io/exclude-from-external-load-balancers
label on control plane nodes. Metallb respects this label and avoids
announcing services from these nodes.

Since we create a single-node kind cluster in the script, the
LoadBalancer IPs for services remain unreachable from the host network.

To resolve this issue, we have two options:
1. Remove the label from nodes, as documented on the kubeadm page.
2. Upgrade metallb version to 0.14.4, which supports ignoring the label
   using the --ignore-exclude-lb flag.

This PR implements option 1, as it’s the simplest solution and aligns with
our existing approach for removing taints.

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>