Relabel bind mounts for SELinux
When SELinux is enabled and enforcing, bind mounts need to be
relabeled to work correctly in many cases. This is configured by
adding the "z" option to --volume definitions (it is not supported
with --mount type=bind).

This adds a relabeling function to run.sh which adds the appropriate
option to --volume definitions, and warns if it finds a --mount
type=bind option. It also converts --mount type=bind declarations to
the corresponding --volume definition.

Signed-off-by: Stephen Kitt <skitt@redhat.com>